@@ -2,10 +2,10 @@ name: tentura_sdk
 description: Dart package for using Tentura server API
 publish_to: none
 
-version: 0.0.1
+version: 0.0.2
 
 environment:
-  sdk: '>=3.5.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 
 dependencies:
   http: ^1.2.2
@@ -13,6 +13,7 @@ dependencies:
   gql_exec: ^1.0.0
   gql_http_link: ^1.1.0
   gql_error_link: ^1.0.0
+  gql_dedupe_link: ^2.0.3
   ed25519_edwards: ^0.3.1
   gql_websocket_link: ^2.0.1
 
