@@ -2,18 +2,18 @@ name: tentura
 description: 'Social network for communities'
 publish_to: 'none'
 
-version: 0.9.2
+version: 0.9.6
 
 environment:
-  sdk: '>=3.5.3 <4.0.0'
-  flutter: ^3.24.5
+  sdk: '>=3.6.0 <4.0.0'
+  flutter: ^3.27.0
 
 dependencies:
   flutter:
     sdk: flutter
   nil: ^1.1.1
   uuid: ^4.4.2
-  intl: ^0.19.0
+  intl: ^0.20.1
   get_it: ^8.0.2
   logger: ^2.5.0
   latlong2: ^0.9.1
@@ -23,18 +23,18 @@ dependencies:
   built_value: ^8.9.2
   flutter_bloc: ^8.1.6
   synchronized: ^3.3.0
-  sentry_flutter: ^8.7.0
+  sentry_flutter: ^8.11.0
   json_annotation: ^4.9.0
   freezed_annotation: ^2.4.4
 
   # Platform related
-  drift: ^2.21.0
+  drift: ^2.22.1
   geocoding: ^3.0.0
-  drift_dev: ^2.21.2
+  drift_dev: ^2.22.1
   geolocator: ^13.0.2
-  share_plus: ^10.1.2
+  share_plus: ^10.1.3
   image_picker: ^1.1.2
-  drift_flutter: ^0.2.1
+  drift_flutter: ^0.2.2
   share_handler: ^0.0.22
   mobile_scanner: ^6.0.2
   flutter_secure_storage: ^9.2.2
@@ -45,12 +45,13 @@ dependencies:
   flutter_map: ^7.0.2
   flutter_svg: ^2.0.16
   google_fonts: ^6.2.1
+  mesh_gradient: ^1.3.8
   vector_graphics: ^1.1.15
-  flutter_chat_ui: ^1.6.15
-  flutter_chat_types: ^3.6.2
+  visibility_detector: ^0.4.0
   flutter_native_splash: ^2.4.3
   # smooth_page_indicator: ^1.2.0
   force_directed_graphview: ^0.6.2
+  scrollable_positioned_list: ^0.3.8
   flutter_map_cancellable_tile_provider: ^3.0.0
 
   # Local
@@ -70,7 +71,7 @@ dev_dependencies:
   build_web_compilers: ^4.0.11
   auto_route_generator: ^9.0.0
   injectable_generator: ^2.6.2
-  flutter_launcher_icons: ^0.14.0
+  flutter_launcher_icons: ^0.14.2
 
 flutter:
   uses-material-design: true
