@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
-import 'package:mobile_scanner/mobile_scanner.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 
 import 'package:tentura/consts.dart';
 import 'package:tentura/app/router/root_router.dart';
+import 'package:tentura/ui/dialog/qr_scan_dialog.dart';
 import 'package:tentura/ui/utils/ui_utils.dart';
 import 'package:tentura/ui/theme.dart';
 
@@ -23,11 +23,7 @@ class App extends StatelessWidget {
     await SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
     ]);
-    if (kIsWeb) {
-      MobileScannerPlatform.instance.setBarcodeLibraryScriptUrl(
-        '/packages/zxing.min.js',
-      );
-    }
+    QRScanDialog.init();
     await configureDependencies();
     FlutterNativeSplash.remove();
     runApp(const App());
@@ -56,21 +52,29 @@ class App extends StatelessWidget {
             ],
             reevaluateListenable: router.reevaluateListenable,
           ),
-          builder: (context, child) => kIsWeb &&
-                  MediaQuery.of(context).orientation == Orientation.landscape
-              ? ColoredBox(
-                  color: Theme.of(context).colorScheme.surfaceBright,
-                  child: Center(
-                    child: ConstrainedBox(
-                      constraints: kWebConstraints,
-                      child: AspectRatio(
-                        aspectRatio: kWebAspectRatio,
-                        child: child,
+          builder: (context, child) {
+            if (child == null) return const SizedBox();
+            final media = MediaQuery.of(context);
+            return MediaQuery(
+              data: media.copyWith(
+                textScaler: TextScaler.noScaling,
+              ),
+              child: kIsWeb && media.orientation == Orientation.landscape
+                  ? ColoredBox(
+                      color: Theme.of(context).colorScheme.surfaceBright,
+                      child: Center(
+                        child: ConstrainedBox(
+                          constraints: kWebConstraints,
+                          child: AspectRatio(
+                            aspectRatio: kWebAspectRatio,
+                            child: child,
+                          ),
+                        ),
                       ),
-                    ),
-                  ),
-                )
-              : child ?? const SizedBox(),
+                    )
+                  : child,
+            );
+          },
         );
       },
     );
