@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 import 'package:tentura/domain/entity/identifiable.dart';
+import 'package:tentura/domain/enum.dart';
 
 part 'chat_message.freezed.dart';
 
@@ -9,12 +10,12 @@ part 'chat_message.freezed.dart';
 class ChatMessage with _$ChatMessage implements Identifiable {
   const factory ChatMessage({
     required String id,
-    required String object,
-    required String subject,
+    required String sender,
     required String content,
+    required String reciever,
     required DateTime createdAt,
     required DateTime updatedAt,
-    @Default(false) bool delivered,
+    @Default(ChatMessageStatus.init) ChatMessageStatus status,
   }) = _ChatMessage;
 
   const ChatMessage._();
@@ -22,8 +23,8 @@ class ChatMessage with _$ChatMessage implements Identifiable {
 
 final emptyMessage = ChatMessage(
   id: '',
-  object: '',
-  subject: '',
+  reciever: '',
+  sender: '',
   content: '',
   createdAt: DateTime.fromMillisecondsSinceEpoch(0),
   updatedAt: DateTime.fromMillisecondsSinceEpoch(0),
