@@ -1,16 +1,17 @@
-import 'package:flutter_chat_types/flutter_chat_types.dart';
-
+import 'package:tentura/domain/entity/profile.dart';
 import 'package:tentura/ui/bloc/state_base.dart';
+
+import '../../domain/entity/chat_message.dart';
 
 part 'chat_state.freezed.dart';
 
 @Freezed(makeCollectionsUnmodifiable: false)
 class ChatState with _$ChatState, StateFetchMixin {
   const factory ChatState({
+    required Profile me,
+    required Profile friend,
     required DateTime cursor,
-    required List<Message> messages,
-    @Default(User(id: '')) User me,
-    @Default(User(id: '')) User friend,
+    required List<ChatMessage> messages,
     @Default(FetchStatus.isSuccess) FetchStatus status,
     Object? error,
   }) = _ChatState;
