@@ -1,17 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:auto_route/auto_route.dart';
-import 'package:flutter_chat_ui/flutter_chat_ui.dart' as chat;
 
 import 'package:tentura/app/router/root_router.dart';
+import 'package:tentura/domain/entity/profile.dart';
+import 'package:tentura/ui/widget/avatar_image.dart';
 import 'package:tentura/ui/widget/deep_back_button.dart';
 import 'package:tentura/ui/widget/linear_pi_active.dart';
 import 'package:tentura/ui/bloc/state_base.dart';
 import 'package:tentura/ui/utils/ui_utils.dart';
 
-import 'package:tentura/features/profile/ui/bloc/profile_cubit.dart';
-import 'package:tentura/features/settings/ui/bloc/settings_cubit.dart';
+import 'package:tentura/features/auth/ui/bloc/auth_cubit.dart';
+import 'package:tentura/features/friends/ui/bloc/friends_cubit.dart';
 
 import '../bloc/chat_cubit.dart';
+import '../bloc/chat_news_cubit.dart';
+import '../dialog/on_chat_clear_dialog.dart';
+import '../widget/chat_list.dart';
 
 @RoutePage()
 class ChatScreen extends StatelessWidget implements AutoRouteWrapper {
@@ -24,60 +28,69 @@ class ChatScreen extends StatelessWidget implements AutoRouteWrapper {
 
   @override
   Widget wrappedRoute(BuildContext context) => BlocProvider(
-        create: (context) {
-          return ChatCubit(
-            me: GetIt.I<ProfileCubit>().state.profile,
-            friendId: id,
-          );
-        },
+        create: (_) => ChatCubit(
+          me: GetIt.I<AuthCubit>().state.currentAccount,
+          friend: GetIt.I<FriendsCubit>().state.friends[id]!,
+          updatesStream: GetIt.I<ChatNewsCubit>().updates,
+        ),
         child: this,
       );
 
   @override
-  Widget build(BuildContext context) {
-    final chatCubit = context.read<ChatCubit>();
-    final chatTheme = switch (GetIt.I<SettingsCubit>().state.themeMode) {
-      ThemeMode.dark => const chat.DarkChatTheme(),
-      ThemeMode.light => const chat.DefaultChatTheme(),
-      ThemeMode.system =>
-        MediaQuery.of(context).platformBrightness == Brightness.light
-            ? const chat.DefaultChatTheme()
-            : const chat.DarkChatTheme(),
-    };
-    return Scaffold(
-      // Header
-      appBar: AppBar(
-        title: BlocSelector<ChatCubit, ChatState, String>(
-          selector: (state) => state.friend.firstName ?? '...',
-          builder: (context, state) => Text(state),
-        ),
-        leading: const DeepBackButton(),
-        bottom: PreferredSize(
-          preferredSize: const Size.fromHeight(4),
-          child: BlocSelector<ChatCubit, ChatState, FetchStatus>(
-            selector: (state) => state.status,
-            builder: (context, status) => status.isLoading
-                ? const LinearPiActive()
-                : const SizedBox(height: 4),
+  Widget build(BuildContext context) => Scaffold(
+        appBar: AppBar(
+          leading: const DeepBackButton(),
+          title: BlocSelector<ChatCubit, ChatState, Profile>(
+            selector: (state) => state.friend,
+            builder: (context, state) => Row(
+              children: [
+                AvatarImage.small(userId: state.imageId),
+                Padding(
+                  padding: kPaddingH,
+                  child: Text(
+                    state.title,
+                    maxLines: 1,
+                    overflow: TextOverflow.ellipsis,
+                  ),
+                ),
+              ],
+            ),
+          ),
+          actions: [
+            PopupMenuButton<void>(
+              itemBuilder: (context) => [
+                PopupMenuItem(
+                  onTap: () async {
+                    if (await OnChatClearDialog.show(context) ?? false) {
+                      if (context.mounted) {
+                        // TBD: remove when implemented
+                        showSnackBar(
+                          context,
+                          text: 'Not implemented yet...',
+                        );
+                        await context.read<ChatCubit>().onChatClear();
+                      }
+                    }
+                  },
+                  child: const Text('Clear chat'),
+                ),
+              ],
+            ),
+          ],
+          bottom: PreferredSize(
+            preferredSize: const Size.fromHeight(4),
+            child: BlocSelector<ChatCubit, ChatState, FetchStatus>(
+              selector: (state) => state.status,
+              builder: (context, status) => status.isLoading
+                  ? const LinearPiActive()
+                  : const SizedBox(height: 4),
+            ),
           ),
         ),
-      ),
-
-      // Chat
-      body: BlocConsumer<ChatCubit, ChatState>(
-        listener: showSnackBarError,
-        listenWhen: (p, c) => c.hasError,
-        buildWhen: (p, c) => c.status.isSuccess,
-        builder: (context, state) => chat.Chat(
-          dateIsUtc: true,
-          theme: chatTheme,
-          onMessageVisibilityChanged: chatCubit.onMessageVisibilityChanged,
-          onSendPressed: chatCubit.onSendPressed,
-          onEndReached: chatCubit.onEndReached,
-          messages: state.messages,
-          user: state.me,
+        body: BlocListener<ChatCubit, ChatState>(
+          listener: showSnackBarError,
+          listenWhen: (p, c) => c.hasError,
+          child: const ChatList(),
         ),
-      ),
-    );
-  }
+      );
 }
