@@ -10,8 +10,13 @@ const kSpacingSmall = 8.0;
 const kSpacingMedium = 16.0;
 const kSpacingLarge = 24.0;
 
+const kPaddingAllS = EdgeInsets.all(kSpacingSmall);
+// TBD: rename to kPaddingAllM
 const kPaddingAll = EdgeInsets.all(kSpacingMedium);
+const kPaddingAllL = EdgeInsets.all(kSpacingLarge);
+// TBD: rename to kPaddingHM
 const kPaddingH = EdgeInsets.symmetric(horizontal: kSpacingMedium);
+// TBD: rename to kPaddingVM
 const kPaddingV = EdgeInsets.symmetric(vertical: kSpacingMedium);
 const kPaddingT = EdgeInsets.only(top: kSpacingMedium);
 
@@ -22,9 +27,14 @@ const kPaddingSmallV = EdgeInsets.symmetric(vertical: kSpacingSmall);
 const kWebConstraints = BoxConstraints(minWidth: 600);
 const kWebAspectRatio = 9 / 16;
 
-final _fYMD = DateFormat.yMd();
-String fYMD(DateTime? dateTime) =>
-    dateTime == null ? '' : _fYMD.format(dateTime);
+final _fmtYMd = DateFormat.yMd();
+final _fmtHm = DateFormat.Hm();
+
+String dateFormatYMD(DateTime? dateTime) =>
+    dateTime == null ? '' : _fmtYMd.format(dateTime);
+
+String timeFormatHm(DateTime? dateTime) =>
+    dateTime == null ? '' : _fmtHm.format(dateTime);
 
 ScaffoldFeatureController<SnackBar, SnackBarClosedReason> showSnackBar(
   BuildContext context, {
