@@ -1,7 +1,9 @@
 import 'dart:isolate';
-import 'package:ferry/ferry.dart' show Client, FetchPolicy, Link, OperationType;
+import 'package:ferry/ferry.dart'
+    show Cache, Client, FetchPolicy, Link, MemoryStore, OperationType;
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql_http_link/gql_http_link.dart';
+import 'package:gql_dedupe_link/gql_dedupe_link.dart';
 import 'package:gql_websocket_link/gql_websocket_link.dart';
 
 import 'auth_link.dart';
@@ -21,7 +23,19 @@ Future<Client> buildClient(
   sendPort!.send(InitMessage(receivePort.sendPort));
   final tokenStream = receivePort.asBroadcastStream();
   return Client(
+    cache: Cache(
+      store: MemoryStore(),
+    ),
+
+    //
+    defaultFetchPolicies: {
+      OperationType.query: FetchPolicy.NoCache,
+    },
+
+    //
     link: Link.from([
+      DedupeLink(),
+
       //
       AuthLink(() async {
         sendPort.send(GetTokenRequest());
@@ -42,6 +56,8 @@ Future<Client> buildClient(
         //
         TransportWebSocketLink(
           TransportWsClientOptions(
+            shouldRetry: (_) => true,
+            log: params.isDebugMode ? print : null,
             connectionParams: () async {
               sendPort.send(GetTokenRequest());
               final token = await tokenStream
@@ -75,10 +91,5 @@ Future<Client> buildClient(
         ),
       ),
     ]),
-
-    //
-    defaultFetchPolicies: {
-      OperationType.query: FetchPolicy.NoCache,
-    },
   );
 }
