@@ -1,7 +1,9 @@
 import 'dart:async';
-import 'package:ferry/ferry.dart' show Client, FetchPolicy, Link, OperationType;
+import 'package:ferry/ferry.dart'
+    show Cache, Client, FetchPolicy, Link, MemoryStore, OperationType;
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql_http_link/gql_http_link.dart';
+import 'package:gql_dedupe_link/gql_dedupe_link.dart';
 import 'package:gql_websocket_link/gql_websocket_link.dart';
 
 import 'auth_link.dart';
@@ -14,7 +16,19 @@ Future<Client> buildClient({
   bool isDebugMode = false,
 }) async =>
     Client(
+      cache: Cache(
+        store: MemoryStore(),
+      ),
+
+      //
+      defaultFetchPolicies: {
+        OperationType.query: FetchPolicy.NoCache,
+      },
+
+      //
       link: Link.from([
+        DedupeLink(),
+
         //
         AuthLink(() => getToken().then((v) => v.value)),
 
@@ -30,6 +44,8 @@ Future<Client> buildClient({
             //
             TransportWebSocketLink(
               TransportWsClientOptions(
+                shouldRetry: (_) => true,
+                log: isDebugMode ? print : null,
                 connectionParams: () async {
                   final token = await getToken();
                   return Future.value({
@@ -58,7 +74,4 @@ Future<Client> buildClient({
               },
             )),
       ]),
-      defaultFetchPolicies: {
-        OperationType.query: FetchPolicy.NoCache,
-      },
     );
